requirements.txt
openai
pdfplumber
PyMuPDF
pandas
python-dotenv
